0.0.6

+++ src/main/java/org/kc7bfi/jflac/io/BitInputStream.java
@@ -1,17 +1,17 @@
 /*
  * libFLAC - Free Lossless Audio Codec library Copyright (C) 2000,2001,2002,2003
  * Josh Coalson
- * 
+ *
  * This library is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Library General Public License as published by the Free
  * Software Foundation; either version 2 of the License, or (at your option) any
  * later version.
- * 
+ *
  * This library is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
  * FOR A PARTICULAR PURPOSE. See the GNU Library General Public License for more
  * details.
- * 
+ *
  * You should have received a copy of the GNU Library General Public License
  * along with this library; if not, write to the Free Software Foundation, Inc.,
  * 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
@@ -29,9 +29,11 @@ import org.kc7bfi.jflac.util.CRC16;
 
 /**
  * Bit-wide input stream.
+ *
  * @author kc7bfi
  */
 public class BitInputStream {
+
     private static final int BITS_PER_BLURB = 8;
     private static final int BITS_PER_BLURB_LOG2 = 3;
     private static final byte BLURB_TOP_BIT_ONE = ((byte) 0x80);
@@ -43,19 +45,20 @@ public class BitInputStream {
     private int getBit = 0;
     private int availBits = 0;
     private int totalBitsRead = 0;
-    
+
     private short readCRC16 = 0;
-    
+
     private InputStream inStream;
-    
+
     /**
      * The constructor.
-     * @param is    The InputStream to read bits from
+     *
+     * @param is The InputStream to read bits from
      */
     public BitInputStream(InputStream is) {
         this.inStream = is;
     }
-    
+
     private int readFromStream() throws IOException {
         // first shift the unconsumed buffer data toward the front as much as possible
         if (getByte > 0 && putByte > getByte) {
@@ -63,22 +66,22 @@ public class BitInputStream {
         }
         putByte -= getByte;
         getByte = 0;
-        
+
         // set the target for reading, taking into account blurb alignment
         // blurb == byte, so no gyrations necessary:
         int bytes = buffer.length - putByte;
-        
+
         // finally, read in some data
         bytes = inStream.read(buffer, putByte, bytes);
         if (bytes <= 0) throw new EOFException();
-        
+
         // now we have to handle partial blurb cases:
         // blurb == byte, so no gyrations necessary:
         putByte += bytes;
         availBits += bytes << 3;
         return bytes;
     }
-    
+
     /**
      * Reset the bit stream.
      */
@@ -88,51 +91,57 @@ public class BitInputStream {
         putByte = 0;
         availBits = 0;
     }
-    
+
     /**
      * Reset the read CRC-16 value.
-     * @param seed  The initial CRC-16 value
+     *
+     * @param seed The initial CRC-16 value
      */
     public void resetReadCRC16(short seed) {
         readCRC16 = seed;
     }
-    
+
     /**
      * return the read CRC-16 value.
-     * @return  The read CRC-16 value
+     *
+     * @return The read CRC-16 value
      */
     public short getReadCRC16() {
         return readCRC16;
     }
-    
+
     /**
      * Test if the Bit Stream consumed bits is byte aligned.
-     * @return  True of bit stream consumed bits is byte aligned
+     *
+     * @return True of bit stream consumed bits is byte aligned
      */
     public boolean isConsumedByteAligned() {
         return ((getBit & 7) == 0);
     }
-    
+
     /**
      * return the number of bits to read to align the byte.
-     * @return  The number of bits to align the byte
+     *
+     * @return The number of bits to align the byte
      */
     public int bitsLeftForByteAlignment() {
         return 8 - (getBit & 7);
     }
-    
+
     /**
      * return the number of bytes left to read.
-     * @return  The number of bytes left to read
+     *
+     * @return The number of bytes left to read
      */
     public int getInputBytesUnconsumed() {
         return availBits >> 3;
-    }   
-        
+    }
+
     /**
      * skip over bits in bit stream without updating CRC.
-     * @param bits  Number of bits to skip
-     * @throws IOException  Thrown if error reading from input stream
+     *
+     * @param bits Number of bits to skip
+     * @throws IOException Thrown if error reading from input stream
      */
     public void skipBitsNoCRC(int bits) throws IOException {
         if (bits == 0) return;
@@ -151,11 +160,12 @@ public class BitInputStream {
             readRawUInt(bits);
         }
     }
-    
+
     /**
      * read a single bit.
-     * @return  The bit
-     * @throws IOException  Thrown if error reading input stream
+     *
+     * @return The bit
+     * @throws IOException Thrown if error reading input stream
      */
     public int readBit() throws IOException {
         while (true) {
@@ -175,14 +185,15 @@ public class BitInputStream {
             }
         }
     }
-    
+
     /**
      * read a bit into an integer value.
-     * The bits of the input integer are shifted left and the 
+     * The bits of the input integer are shifted left and the
      * read bit is placed into bit 0.
-     * @param val   The integer to shift and add read bit
-     * @return      The updated integer value
-     * @throws IOException  Thrown if error reading input stream
+     *
+     * @param val The integer to shift and add read bit
+     * @return The updated integer value
+     * @throws IOException Thrown if error reading input stream
      */
     public int readBitToInt(int val) throws IOException {
         while (true) {
@@ -203,15 +214,16 @@ public class BitInputStream {
             }
         }
     }
-    
+
     /**
      * peek at the next bit and add it to the input integer.
-     * The bits of the input integer are shifted left and the 
+     * The bits of the input integer are shifted left and the
      * read bit is placed into bit 0.
-     * @param val   The input integer
-     * @param bit   The bit to peek at
-     * @return      The updated integer value
-     * @throws IOException  Thrown if error reading input stream
+     *
+     * @param val The input integer
+     * @param bit The bit to peek at
+     * @return The updated integer value
+     * @throws IOException Thrown if error reading input stream
      */
     public int peekBitToInt(int val, int bit) throws IOException {
         while (true) {
@@ -229,14 +241,15 @@ public class BitInputStream {
             }
         }
     }
-    
+
     /**
      * read a bit into a long value.
-     * The bits of the input long are shifted left and the 
+     * The bits of the input long are shifted left and the
      * read bit is placed into bit 0.
-     * @param val   The long to shift and add read bit
-     * @return      The updated long value
-     * @throws IOException  Thrown if error reading input stream
+     *
+     * @param val The long to shift and add read bit
+     * @return The updated long value
+     * @throws IOException Thrown if error reading input stream
      */
     public long readBitToLong(long val) throws IOException {
         while (true) {
@@ -257,12 +270,13 @@ public class BitInputStream {
             }
         }
     }
-    
+
     /**
      * read bits into an unsigned integer.
-     * @param bits  The number of bits to read
-     * @return      The bits as an unsigned integer
-     * @throws IOException  Thrown if error reading input stream
+     *
+     * @param bits The number of bits to read
+     * @return The bits as an unsigned integer
+     * @throws IOException Thrown if error reading input stream
      */
     public int readRawUInt(int bits) throws IOException {
         int val = 0;
@@ -271,12 +285,13 @@ public class BitInputStream {
         }
         return val;
     }
-    
+
     /**
      * peek at bits into an unsigned integer without advancing the input stream.
-     * @param bits  The number of bits to read
-     * @return      The bits as an unsigned integer
-     * @throws IOException  Thrown if error reading input stream
+     *
+     * @param bits The number of bits to read
+     * @return The bits as an unsigned integer
+     * @throws IOException Thrown if error reading input stream
      */
     public int peekRawUInt(int bits) throws IOException {
         int val = 0;
@@ -285,20 +300,23 @@ public class BitInputStream {
         }
         return val;
     }
-    
+
     /**
      * read bits into a signed integer.
-     * @param bits  The number of bits to read
-     * @return      The bits as a signed integer
-     * @throws IOException  Thrown if error reading input stream
+     *
+     * @param bits The number of bits to read
+     * @return The bits as a signed integer
+     * @throws IOException Thrown if error reading input stream
      */
-    public int readRawInt(int bits) throws IOException { 
-        if (bits == 0) { return 0; }
+    public int readRawInt(int bits) throws IOException {
+        if (bits == 0) {
+            return 0;
+        }
         int uval = 0;
         for (int i = 0; i < bits; i++) {
             uval = readBitToInt(uval);
         }
-        
+
         // fix the sign
         int val;
         int bitsToleft = 32 - bits;
@@ -311,12 +329,13 @@ public class BitInputStream {
         }
         return val;
     }
-    
+
     /**
      * read bits into an unsigned long.
-     * @param bits  The number of bits to read
-     * @return      The bits as an unsigned long
-     * @throws IOException  Thrown if error reading input stream
+     *
+     * @param bits The number of bits to read
+     * @return The bits as an unsigned long
+     * @throws IOException Thrown if error reading input stream
      */
     public long readRawULong(int bits) throws IOException {
         long val = 0;
@@ -325,11 +344,12 @@ public class BitInputStream {
         }
         return val;
     }
-    
+
     /**
      * read bits into an unsigned little endian integer.
-     * @return      The bits as an unsigned integer
-     * @throws IOException  Thrown if error reading input stream
+     *
+     * @return The bits as an unsigned integer
+     * @throws IOException Thrown if error reading input stream
      */
     public int readRawIntLittleEndian() throws IOException {
         int x32 = readRawUInt(8);
@@ -341,15 +361,16 @@ public class BitInputStream {
         x32 |= (x8 << 24);
         return x32;
     }
-    
+
     /**
      * Read a block of bytes (aligned) without updating the CRC value.
+     *
      * @param val   The array to receive the bytes. If null, no bytes are returned
      * @param nvals The number of bytes to read
-     * @throws IOException  Thrown if error reading input stream
+     * @throws IOException Thrown if error reading input stream
      */
     public void readByteBlockAlignedNoCRC(byte[] val, int nvals) throws IOException {
-        int destlength = nvals;        
+        int destlength = nvals;
         while (nvals > 0) {
             int chunk = Math.min(nvals, putByte - getByte);
             if (chunk <= 0) {
@@ -364,11 +385,12 @@ public class BitInputStream {
             }
         }
     }
-    
+
     /**
      * Read and count the number of zero bits.
-     * @return  The number of zero bits read
-     * @throws IOException  Thrown if error reading input stream
+     *
+     * @return The number of zero bits read
+     * @throws IOException Thrown if error reading input stream
      */
     public int readUnaryUnsigned() throws IOException {
         int val = 0;
@@ -379,14 +401,15 @@ public class BitInputStream {
         }
         return val;
     }
-    
+
     /**
      * Read a Rice Signal Block.
-     * @param vals  The values to be returned
-     * @param pos   The starting position in the vals array
-     * @param nvals The number of values to return
+     *
+     * @param vals      The values to be returned
+     * @param pos       The starting position in the vals array
+     * @param nvals     The number of values to return
      * @param parameter The Rice parameter
-     * @throws IOException  On read error
+     * @throws IOException On read error
      */
     public void readRiceSignedBlock(int[] vals, int pos, int nvals, int parameter) throws IOException {
         int j, valI = 0;
@@ -395,9 +418,9 @@ public class BitInputStream {
         int state = 0; // 0 = getting unary MSBs, 1 = getting binary LSBs
         if (nvals == 0) return;
         int i = getByte;
-        
+
         long startBits = getByte * 8L + getBit;
-        
+
         // We unroll the main loop to take care of partially consumed blurbs here.
         if (getBit > 0) {
             saveBlurb = blurb = buffer[i];
@@ -409,7 +432,7 @@ public class BitInputStream {
                         for (j = 0; (blurb & BLURB_TOP_BIT_ONE) == 0; j++)
                             blurb <<= 1;
                         msbs += j;
-                        
+
                         // dispose of the unary end bit
                         blurb <<= 1;
                         j++;
@@ -482,7 +505,7 @@ public class BitInputStream {
 //            totalConsumedBits = (i << BITS_PER_BLURB_LOG2) | cbits;
 //            totalBitsRead += (BITS_PER_BLURB) | cbits;
         }
-        
+
         // Now that we are blurb-aligned the logic is slightly simpler
         while (valI < nvals) {
             for (; i < putByte && valI < nvals; i++) {
@@ -568,7 +591,7 @@ public class BitInputStream {
                 long endBits = getByte * 8L + getBit;
 //                System.out.println("SE0 " + startBits + " " + endBits);
                 totalBitsRead += (int) (endBits - startBits);
-                availBits -=  endBits - startBits;
+                availBits -= (int) (endBits - startBits);
                 readFromStream();
                 // these must be zero because we can only get here if we got to
                 // the end of the buffer
@@ -580,19 +603,20 @@ public class BitInputStream {
         long endBits = getByte * 8L + getBit;
 //        System.out.println("SE1 " + startBits + " " + endBits);
         totalBitsRead += (int) (endBits - startBits);
-        availBits -= endBits - startBits;
+        availBits -= (int) (endBits - startBits);
     }
-    
+
     /**
      * read UTF8 integer.
      * on return, if *val == 0xffffffff then the utf-8 sequence was invalid, but
      * the return value will be true
-     * @param raw   The raw bytes read (output). If null, no bytes are returned
-     * @return      The integer read
-     * @throws IOException  Thrown if error reading input stream
+     *
+     * @param raw The raw bytes read (output). If null, no bytes are returned
+     * @return The integer read
+     * @throws IOException Thrown if error reading input stream
      */
     public int readUTF8Int(ByteData raw) throws IOException {
- 
+
         int v;
         int x;
         int i;
@@ -632,14 +656,15 @@ public class BitInputStream {
         }
         return v;
     }
-    
+
     /**
      * read UTF long.
      * on return, if *val == 0xffffffffffffffff then the utf-8 sequence was
      * invalid, but the return value will be true
-     * @param raw   The raw bytes read (output). If null, no bytes are returned
-     * @return      The long read
-     * @throws IOException  Thrown if error reading input stream
+     *
+     * @param raw The raw bytes read (output). If null, no bytes are returned
+     * @return The long read
+     * @throws IOException Thrown if error reading input stream
      */
     public long readUTF8Long(ByteData raw) throws IOException {
         long v;
@@ -685,9 +710,10 @@ public class BitInputStream {
         }
         return v;
     }
-    
+
     /**
      * Total Blurbs read.
+     *
      * @return Returns the total blurbs read.
      */
     public int getTotalBytesRead() {
